+++ src/components/SideBar/index.css
@@ -54,26 +54,6 @@
   margin-left: 15px;
 }
 
-/* search file css */
-.search-bar input{
-  width:25%;
-  }
-  .search-bar{
-    display: flex;
-    margin: 100px auto;
-    text-align: center;
-    align-items: center;
-    justify-content: center;
-  }
-  .search-bar button{
-    border: none;
-    background: #0685f5;
-    color: white;
-    text-align: center;
-    padding: 10px;
-    border-radius: 0px 20px ;
-    margin-left: 2px;
-  }
 /* media query */
 
 @media screen and (max-width: 1200px) {
